+++ internal/command/sync.go
@@ -107,6 +107,9 @@ func (c *SyncCommand) executeSave(args []string, stdout, stderr io.Writer) error
 		}
 		return err
 	}
+	if fs.NArg() > 0 {
+		return fmt.Errorf("unexpected arguments for save: %v", fs.Args())
+	}
 
 	if title == "" {
 		_, _ = fmt.Fprintln(stderr, "Error: --title is required")
@@ -175,6 +178,9 @@ func (c *SyncCommand) executeList(args []string, stdout, stderr io.Writer) error
 		}
 		return err
 	}
+	if fs.NArg() > 0 {
+		return fmt.Errorf("unexpected arguments for list: %v", fs.Args())
+	}
 
 	dir, err := c.notebooksDir()
 	if err != nil {
@@ -230,6 +236,9 @@ func (c *SyncCommand) executeInit(args []string, stdout, stderr io.Writer) error
 		return err
 	}
 
+	if fs.NArg() > 1 {
+		return fmt.Errorf("unexpected arguments for init: %v", fs.Args()[1:])
+	}
 	repoURL := ""
 	if fs.NArg() > 0 {
 		repoURL = fs.Arg(0)
@@ -263,6 +272,9 @@ func (c *SyncCommand) executeInit(args []string, stdout, stderr io.Writer) error
 
 // executePush stages all changes, commits, and pushes to the remote.
 func (c *SyncCommand) executePush(args []string, stdout, stderr io.Writer) error {
+	if len(args) > 0 {
+		return fmt.Errorf("unexpected arguments for push: %v", args)
+	}
 	root, err := c.syncRoot()
 	if err != nil {
 		return err
@@ -299,6 +311,9 @@ func (c *SyncCommand) executePush(args []string, stdout, stderr io.Writer) error
 
 // executePull fetches and merges remote changes, or clones if not initialized.
 func (c *SyncCommand) executePull(args []string, stdout, stderr io.Writer) error {
+	if len(args) > 0 {
+		return fmt.Errorf("unexpected arguments for pull: %v", args)
+	}
 	root, err := c.syncRoot()
 	if err != nil {
 		return err